pyproject.toml
[project]
name = "flower-classification"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.129.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "pillow>=12.1.1",
    "python-multipart>=0.0.22",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.0",
    "seaborn>=0.13.2",
    "tensorflow>=2.20.0",
    "tensorflow-datasets>=4.9.9",
    "uvicorn>=0.41.0",
]
